Refresh_book_list

+++ Library Assistant/src/library/assistant/ui/listbook/BookListController.java
@@ -5,6 +5,7 @@
  */
 package library.assistant.ui.listbook;
 
+import java.io.IOException;
 import java.net.URL;
 import java.sql.ResultSet;
 import java.sql.SQLException;
@@ -18,7 +19,10 @@ import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
+import javafx.fxml.FXMLLoader;
 import javafx.fxml.Initializable;
+import javafx.scene.Parent;
+import javafx.scene.Scene;
 import javafx.scene.control.Alert;
 import javafx.scene.control.Alert.AlertType;
 import javafx.scene.control.ButtonType;
@@ -26,8 +30,14 @@ import javafx.scene.control.TableColumn;
 import javafx.scene.control.TableView;
 import javafx.scene.control.cell.PropertyValueFactory;
 import javafx.scene.layout.AnchorPane;
+import javafx.stage.Stage;
+import javafx.stage.StageStyle;
 import library.assistant.alert.AlertMaker;
 import library.assistant.database.DatabaseHandler;
+import library.assistant.ui.addbook.BookAddController;
+import library.assistant.ui.main.Main;
+import library.assistant.ui.main.MainController;
+import library.assistant.util.LibraryAssistantUtil;
 
 /**
  * FXML Controller class
@@ -55,6 +65,7 @@ public class BookListController implements Initializable {
 
     /**
      * Initializes the controller class.
+     *
      * @param url
      * @param rb
      */
@@ -75,6 +86,7 @@ public class BookListController implements Initializable {
 
     private void loadData() {
         try {
+            list.clear();
             DatabaseHandler handler = DatabaseHandler.getInstance();
             String query = "SELECT * FROM BOOK";
             ResultSet rs = handler.execQuery(query);
@@ -84,37 +96,90 @@ public class BookListController implements Initializable {
                 String gauthor = rs.getString("author");
                 String gpublisher = rs.getString("publisher");
                 Boolean gavailability = rs.getBoolean("isAvail");
-                list.add(new Book(gtitle,gid,gauthor,gpublisher,gavailability)); 
+                list.add(new Book(gtitle, gid, gauthor, gpublisher, gavailability));
             }
 
         } catch (SQLException ex) {
             Logger.getLogger(BookListController.class.getName()).log(Level.SEVERE, null, ex);
         }
-          tableView.setItems(list);   //   tableView.getItems().setAll(list);
+        tableView.setItems(list);   //   tableView.getItems().setAll(list);
     }
 
     @FXML
     private void context_menu_delete(ActionEvent event) throws SQLException {
         Book bookselected = tableView.getSelectionModel().getSelectedItem();
-        if(bookselected==null) {
+        if (bookselected == null) {
             AlertMaker.showErrorAlert("Error", "No book Selected.");
-        return;
+            return;
         }
-        Alert alert = new Alert(AlertType.CONFIRMATION);
-        alert.setContentText("Are you sure want to delete the "+bookselected.getTitle()+" ?");
-        Optional<ButtonType> answer = alert.showAndWait();
-        if(answer.get()==ButtonType.OK) {
-            Boolean result  = DatabaseHandler.getInstance().deleteBook(bookselected);
-            if (result) {
-                AlertMaker.showSimpleAlert("Success", "Book has been deleted successfully.");
-                list.remove(bookselected);
+
+        if (!DatabaseHandler.getInstance().isBookAlreadyIssued(bookselected)) {
+            System.out.println(DatabaseHandler.getInstance().isBookAlreadyIssued(bookselected));
+            Alert alert = new Alert(AlertType.CONFIRMATION);
+            alert.setContentText("Are you sure want to delete the " + bookselected.getTitle() + " ?");
+            Optional<ButtonType> answer = alert.showAndWait();
+            if (answer.get() == ButtonType.OK) {
+                Boolean result = DatabaseHandler.getInstance().deleteBook(bookselected);
+                if (result) {
+                    AlertMaker.showSimpleAlert("Success", "Book has been deleted successfully.");
+                    list.remove(bookselected);
+                } else {
+                    AlertMaker.showErrorAlert("Failed", "Failed to delete Book");
+                }
             } else {
-                AlertMaker.showErrorAlert("Failed", "Failed to delete Book");
+                AlertMaker.showSimpleAlert("Cancelled", "You cancelled the book deletion");
             }
-        }else {
-            AlertMaker.showSimpleAlert("Cancelled", "You cancelled the book deletion");
+
+//        } else {
+//            Alert alert = new Alert(AlertType.ERROR);
+//            alert.setHeaderText(null);
+//            alert.setContentText("Book already issued, Can't be deleted.");
+//            alert.showAndWait();
+//        }
+//       
+            //  System.out.println("already issued");
+        } else {
+            AlertMaker.showErrorAlert("Can't issue ", "Can't Book is already issued to someone");
+
         }
-        
+    }
+
+    @FXML
+    private void context_menu_edit(ActionEvent event) throws IOException {
+        Book selectedforedit = tableView.getSelectionModel().getSelectedItem();
+        if (selectedforedit == null) {
+            AlertMaker.showErrorAlert("Error", "No book selected.");
+            return;
+        }
+        FXMLLoader loader = new FXMLLoader(getClass().getResource("/library/assistant/ui/addbook/add_book.fxml"));
+        Parent parent = loader.load();
+
+        BookAddController controller = (BookAddController) loader.getController();
+        controller.infalateUI(selectedforedit);
+
+        Stage stage = new Stage(StageStyle.DECORATED);
+        stage.setTitle("Edit Book");
+        LibraryAssistantUtil.setStageIcon(stage);
+        stage.setScene(new Scene(parent));
+        stage.show();
+        stage.setOnCloseRequest(e -> {
+            System.out.println("refreshinng data");
+            context_menu_refresh(new ActionEvent());
+        });
+
+//        
+//        MainController controller = new MainController();
+//        BookAddController bac = new BookAddController();
+//        bac.infalateUI(selectedforedit);
+//        
+//        controller.loadWindow("/library/assistant/ui/addbook/add_book.fxml", "Edit Book");
+//        System.out.println("loaded eddit winodqw");
+//       
+    }
+
+    @FXML
+    private void context_menu_refresh(ActionEvent actionEvent) {
+        loadData();
     }
 
     public static class Book {
@@ -125,7 +190,7 @@ public class BookListController implements Initializable {
         private final SimpleStringProperty publisher;
         private final SimpleBooleanProperty availability;
 
-        Book(String title, String id, String author, String publisher, Boolean availability) {
+        public Book(String title, String id, String author, String publisher, Boolean availability) {
             this.title = new SimpleStringProperty(title);
             this.id = new SimpleStringProperty(id);
             this.author = new SimpleStringProperty(author);
